requirements.txt
Flask
reportlab
Pillow
qrcode
pywebview
werkzeug